# docs/swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"Returns a greeting string.","content":{"application/json":{"schema":{"type":"string","example":"Hello World!"}}}}},"summary":"Get a greeting message","tags":["Application"]}},"/health":{"get":{"operationId":"AppController_getHealth","parameters":[],"responses":{"200":{"description":"Returns health status, uptime, and timestamp of the server.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthStatusDto"}}}}},"summary":"Get application health status","tags":["Application"]}},"/products":{"get":{"operationId":"ProductsController_findAll","parameters":[{"name":"category","required":false,"in":"query","description":"Filter by category","schema":{"type":"string","enum":["GROCERY","BEVERAGE","PERSONAL_CARE","HOUSEHOLD","ELECTRONICS","CLOTHING","TOYS","STATIONERY","OTHER"]}},{"name":"isActive","required":false,"in":"query","description":"Filter by active status (true/false)","schema":{"type":"string"}},{"name":"name","required":false,"in":"query","description":"Search by product name (ILIKE)","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","description":"Page number (default: 1)","schema":{"default":1,"type":"number"}},{"name":"perPage","required":false,"in":"query","description":"Items per page (default: 10)","schema":{"default":10,"type":"number"}}],"responses":{"200":{"description":"List of products with pagination metadata","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedProductSummaryDto"}}}}},"summary":"List all products with filters and pagination","tags":["Products"]},"post":{"operationId":"ProductsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":"Created product details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponseDto"}}}}},"summary":"Create a new product","tags":["Products"]}},"/products/{id}":{"get":{"operationId":"ProductsController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"Product UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"Complete product details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponseDto"}}}}},"summary":"Get product details by ID","tags":["Products"]},"patch":{"operationId":"ProductsController_update","parameters":[{"name":"id","required":true,"in":"path","description":"Product UUID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":"Updated product details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponseDto"}}}}},"summary":"Update a product by ID","tags":["Products"]},"delete":{"operationId":"ProductsController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"Product UUID","schema":{"type":"string"}}],"responses":{"204":{"description":"Product successfully deleted"}},"summary":"Delete a product by ID","tags":["Products"]}}},"info":{"title":"Products microservice - Fruto Infinito","description":"","version":"1.0.0","contact":{"name":"Fruto Infinito Dev Team","url":"https://frutoinfinito.com","email":"support@frutoinfinito.com"},"license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"}},"tags":[],"servers":[],"components":{"schemas":{"HealthStatusDto":{"type":"object","properties":{"status":{"type":"string","example":"ok","description":"Application health status"},"uptime":{"type":"string","example":"01h 23m 45s","description":"Formatted application uptime"},"timestamp":{"type":"string","example":"07/08/2025, 14:25:30","description":"Current timestamp of the server"}},"required":["status","uptime","timestamp"]},"ProductSummaryDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique product identifier","example":"b28d24b3-e5ce-4e75-b0ec-328ffb611ae4"},"name":{"type":"string","description":"Name of the product","example":"Organic Green Tea"},"price":{"type":"number","description":"Unit price of the product","example":9.99},"category":{"type":"string","description":"Product category","enum":["GROCERY","BEVERAGE","PERSONAL_CARE","HOUSEHOLD","ELECTRONICS","CLOTHING","TOYS","STATIONERY","OTHER"],"example":"BEVERAGE"}},"required":["id","name","price","category"]},"MetaDto":{"type":"object","properties":{"total":{"type":"number","description":"Total number of products available","example":125},"page":{"type":"number","description":"Current page number","example":2},"perPage":{"type":"number","description":"Number of products returned per page","example":10}},"required":["total","page","perPage"]},"PaginatedProductSummaryDto":{"type":"object","properties":{"products":{"description":"List of products for the current page","type":"array","items":{"$ref":"#/components/schemas/ProductSummaryDto"}},"meta":{"description":"Pagination metadata","allOf":[{"$ref":"#/components/schemas/MetaDto"}]}},"required":["products","meta"]},"ProductResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique product identifier","example":"b28d24b3-e5ce-4e75-b0ec-328ffb611ae4"},"name":{"type":"string","description":"Name of the product","example":"Organic Green Tea"},"description":{"type":"string","description":"Detailed description of the product","example":"A pack of 20 organic green tea bags."},"price":{"type":"number","description":"Unit price of the product","example":9.99},"category":{"type":"string","description":"Product category","enum":["GROCERY","BEVERAGE","PERSONAL_CARE","HOUSEHOLD","ELECTRONICS","CLOTHING","TOYS","STATIONERY","OTHER"],"example":"BEVERAGE"},"isActive":{"type":"boolean","description":"Whether the product is active","example":true},"createdAt":{"format":"date-time","type":"string","description":"Creation timestamp","example":"2024-07-01T14:23:00.000Z"},"updatedAt":{"format":"date-time","type":"string","description":"Last update timestamp","example":"2024-07-03T11:05:00.000Z"}},"required":["id","name","description","price","category","isActive","createdAt","updatedAt"]},"CreateProductDto":{"type":"object","properties":{"name":{"type":"string","description":"Name of the product","example":"Organic Green Tea"},"description":{"type":"string","description":"Detailed description of the product","example":"A pack of 20 organic green tea bags with no additives."},"price":{"type":"number","description":"Unit price of the product","example":9.99},"category":{"type":"string","description":"Category the product belongs to","enum":["GROCERY","BEVERAGE","PERSONAL_CARE","HOUSEHOLD","ELECTRONICS","CLOTHING","TOYS","STATIONERY","OTHER"],"example":"BEVERAGE"}},"required":["name","description","price","category"]},"UpdateProductDto":{"type":"object","properties":{"name":{"type":"string","description":"Name of the product","example":"Organic Green Tea"},"description":{"type":"string","description":"Detailed description of the product","example":"A pack of 20 organic green tea bags with no additives."},"price":{"type":"number","description":"Unit price of the product","example":9.99},"category":{"type":"string","description":"Category the product belongs to","enum":["GROCERY","BEVERAGE","PERSONAL_CARE","HOUSEHOLD","ELECTRONICS","CLOTHING","TOYS","STATIONERY","OTHER"],"example":"BEVERAGE"}}}}}}